Improve log message when mapping keysyms
It was a poor choice of words to call these keysyms "unknown" as they
are likely perfectly normal keysyms, they just cannot be found in the
currently loaded keyboard layout. This log message has confused users,
so let's get rid of the "unknown" part.

We already use the verb "add" heavily in the other log messages, so
let's keep that rather than switch to some for of "map".