Remove Logger_File mutex
This is not providing protection for all of the logging, so it gives a
false sense of security. We're also not making use of this as we have
barely any threading.

The exception is the WinVNC code, which is very thread heavy. But it is
unmaintained and will not be a blocker for this.