Avoid fetching details of disabled pointer
Certain versions of the xserver code will crash if you access the
details of a disabled pointer. The details are also not up to date when
it is disabled, so it is likely concepually wrong to ignore it at this
point anyway.

Issue originally discovered and diagnosed by Mark Mielke.