Move all remote resize details to remoteResize()
This complex logic is difficult to understand when it is spread out.
Change DesktopWindow::remoteResize() to be a request to resize instead,
and let it decide fully if and when it will honour that request.

Note that this slightly changes the behaviour of -DesktopSize. It will
no longer override the -RemoteResize setting.