Add SELinux policy rules for non-existing directories
We have policy rules for Tigervnc specific directories, but we don't
have rules for the parent directories, which might need to be created
as well.

E.g. we have a rule for ~/.local/share/tigervnc, but we don't have a
rule for ~/.local/share, which might need to be created first with
different SELinux context.