Run Google Test unit tests in GitHub actions
Give us constant feedback that nothing breaks.

Note that we need a separate workflow to present the test results.
Workflows that are run for a pull request aren't allowed to modify the
pull request to add more checks (for security reasons). Instead, we save
the test results as an artifact and add a second workflow that only
presents those results.