Keep input methods constantly disabled
It is unfortunately not reliable to turn them on and off with focus, at
least not on macOS. We lose focus when the layout selector is opened.
Which means the disabled layouts are available to the user to chose, and
the result is a bit random.

We currently don't have a practical need to re-enable input methods once
the session is running as we don't have any input fields at this point.
So for now, simply disable the input methods constantly whilst connected
to work around the issue.