Use generic hot key combo for client control
We want to be able to define more keyboard combinations that allows the
user to control the client instead of sending the keys to the server.
Instead of adding a never ending stream of random keys that just
confuses things, we'll define a set of modifiers that are the base for
all client control commands.