Regrab keyboard on temporary focus loss
If we lose focus whilst the keyboard grab is active, then it should mean
we didn't lose it the way you normally do using the keyboard (e.g.
Alt+Tab). Common scenarios are opening the context menu, or clicking on
another monitor or application.

In these cases, the user likely wants to keep the same state, i.e.
grabbed, when they return to the viewer window.