Add view-only w0vncserver
This commits adds a view-only VNC server that works on local Wayland
sessions that have the ScreenCast portal implemented.

It uses PipeWire to access the desktop's screen content.

The RemoteDesktop class currently only uses the ScreenCast portal. In a
future commit, it will also use the RemoteDesktop portal. To avoid
re-naming the class later, let's just call it RemoteDesktop now.