w0vncserver: Add local cursor support
By setting 'cursor_mode' to METADATA, PipeWire will send cursor metadata
in the same stream as screen content instead of having it embedded in
the screen buffer.

I saw some issues with enabling cursor metadata with mutter where some
frames would be skipped. And on KDE, the cursor is rendered as part of
the video stream even with 'cursor_mode' set to METADATA. Let's leave
this feature as an option that is disabled by default until things
change.