PortalDesktop: Enable using restore_token
Restore tokens can be used to save/restore a DBus session. This makes it
possible to re-use a previous session, bypassing the remote desktop
dialog that pops up.

In this commit, a new parameter "AskDisplayChoice" is added where
users can choose between three options:

  1. "Always" - The remote desktop dialog will always pop up and restore
     tokens are not used. This is identical to the behaviour prior to
     this commit.

  2. "Once" - Only display the remote desktop dialogue the first time a
     user connects. Subsequent connections will re-use the same session
     that was created when the first user connected. The session is
     valid as long as the server is not closed. Users have to restart
     the server to get the dialogue again.

  3. "Never" - Display the remote desktop dialogue once, and remember
     the choice even after restarts. The restore token is stored to disk
     and can be loaded when the RemoteDesktop session starts.

The default option right now is "Once", which should be helpful for most
users. Having "Never" as the default can be problematic as it might not
be obvious to users how to "reset" their display choice.

To clear the previously stored choice, utility program called
'w0vncserver-forget' can be used. Users can simply call this program to
forget their display choice.