Wrapper for allowing multiple Fl::event_dispatch
Sometimes FLTK widgets have to listen to events on the parent window.
This can be done using Fl::event_dispatch(), but each window can only
have one such handler. But we must have the possilibity of multiple
instances of a widget within a window. A general wrapper for
Fl::event_dispatch() was written for this purpose.