Searchable server history suggestions
This adds a new widget for displaying suggestions in a 'searchable' way,
built into an input widget. It is used in ServerDialog to display the
history of VNC servers instead of the Fl_Input_Choice widget.

The server name input field will show a drop-down list below if there
are any historical VNC servers. At most 4 suggestions are shown at the
same time. The following three actions trigger the suggestions to show:

- Typing in the input field
- Clicking the input field
- Pressing the down arrow key while focused

The suggestion drop-down can be closed by a click outside or by pressing
ESC (utilizes event_dispatch_handler).

The suggestions match the user input and can be selected by a click or
by pressing the enter key. Suggestions can be removed by clicking the x
button. Suggestions can be cycled by pressing the up and down arrow
keys. Suggestion matches are emphasized with a bold font, and separate
words can be used to search for partial matches. Matching is
case-insensitive because DNS names are case-insensitive.

The list can display both matches and historical suggestions at the same
time. The list is sorted and shows the most relevant matches first, with
historical suggestions below.