Detect dropped frames in PipeWirePixelBuffer
If we don't process the events coming from PipeWire quick enough, we
might skip processing some frames. The effect we saw from this was that
the damage tracking would be incorrect, and the client would get sent
invalid frames.

PipeWire sends us a sequence with each buffer which we can use to detect
skipped/dropped events. When that happens, we can no longer rely on our
damage tracking and have to mark the entire display as dirty.