PortalDesktop: Remember clipboard preference
Just as with the restore token, we need to kep track of the clipboard
preference for a given session so that we know if clipboard should be
enabled or not the next time w0vncserver starts.

It seems you have to explicitly call Clipboard.RequestClipboard before
starting the session, even if it was requested and enabled for the
previously stored session.