PortalDesktop: Add Clipboard support
This commit adds clipboard support to PortalDesktop through the
Clipboard portal.

The Clipboard portal can be integrated with the Remote Desktop portal to
make it possible to do clipboard handling from the remote desktop.

It uses MIME types to advertise available formats, which means we can't
rely on X11 targets such as TEXT/UTF8_STRING etc.

We currently support the `text/plain` and `text/plain;charset=utf-8`
MIME types, which are widely used.