Transform hooked key events separately
The low-level keyboard hooks is latency sensitive, and is executed on
its own thread. Let's do as little work as we possibly can there, and
instead use a more specific hook to massage the messages as needed.

This later hook will be executed as messages are grabbed from the queue.
So there is no risk for races or other odd behaviour.