Use correct key code for key release on Windows
We're now tracking the raw system key codes, not the translated ones we
send to the server. These are mostly the same on Windows, but there are
a few exceptions which results in keys getting stuck.

This was overlooked in the refactoring in 7e7e050.

(cherry picked from commit 6d2f2da1c0914c3b15f2b6e2511ef7397dfeeb66)