Avoid warning about unused checkedBytes
It's only used in debug builds, so clang will complain that it is unused
in a release build. We don't want to #ifdef the attribute, though, as
that changes the layout of the object, and developers might be toggling
debug and release builds in a way that would make that unsafe.